requirements.txt
requests==2.16.5
python-telegram-bot==12.0.0b1
